fix(overlay): actually remove the circular dependency
